add settings with properties
